requirements.txt
Flask==3.0.3
gunicorn==22.0.0
markdown==3.6
python-slugify==8.0.4
